added indexes. changed score in history to int instead of bigint

+++ QuizWebsite/src/main/java/org/ja/servlet/GradeAllQuestionsServlet.java
@@ -79,7 +79,7 @@ public class GradeAllQuestionsServlet extends HttpServlet {
             completionTime = quiz.getTimeInMinutes();
 
         session.setAttribute("time-spent-in-minutes", completionTime);
-        long totalScore = grades.stream().mapToInt(Integer::intValue).sum();
+        int totalScore = grades.stream().mapToInt(Integer::intValue).sum();
 
         Timestamp completionDate = new Timestamp(endTime);
